+++ spec/examples/schema/table_synchronizer_spec.rb
@@ -2,7 +2,10 @@
 require File.expand_path('../../spec_helper', __FILE__)
 
 describe Cequel::Schema::TableSynchronizer do
-  let(:table_name) { :"posts_#{SecureRandom.hex(4)}" }
+  let(:table_name) { |ex|
+    ent = SecureRandom.hex(4)
+    unique_table_name("posts_#{ent}", ex)
+  }
 
   let(:table) { cequel.schema.read_table(table_name) }
 
@@ -41,55 +44,108 @@ describe Cequel::Schema::TableSynchronizer do
 
     after { cequel.schema.drop_table(table_name) }
 
-    context 'with valid changes' do
-
-      before do
-        cequel.schema.sync_table table_name do
-          key :blog_subdomain, :text
-          key :post_permalink, :text
-          column :title, :ascii
-          column :body, :text
-          column :primary_author_id, :uuid, :index => true
-          column :created_at, :timestamp, :index => true
-          column :published_at, :timestamp
-          set :author_names, :text
-          list :categories, :text
-          with :comment, 'Test Table 2.0'
-        end
+    it 'should rename cluster keys' do
+      cequel.schema.sync_table table_name do
+        key :blog_subdomain, :text
+        key :post_permalink, :text  # renamed
+        column :title, :ascii, :index => true
+        column :body, :ascii
+        column :created_at, :timestamp
+        set :author_names, :text
+        with :comment, 'Test Table'
       end
 
-      it 'should rename keys' do
-        expect(table.clustering_columns.first.name).to eq(:post_permalink)
+      expect(table.clustering_columns.first.name).to eq(:post_permalink)
+    end
+
+    it 'should add new columns' do |ex|
+      cequel.schema.sync_table table_name do
+        key :blog_subdomain, :text
+        key :permalink, :text
+        column :title, :ascii, :index => true
+        column :body, :ascii
+        column :created_at, :timestamp
+        set :author_names, :text
+        with :comment, 'Test Table - #{ex.description}'
+
+        column :published_at, :timestamp # new
       end
 
-      it 'should add new columns' do
-        expect(table.column(:published_at).type).to eq(Cequel::Type[:timestamp])
+      expect(table.column(:published_at).type).to eq(Cequel::Type[:timestamp])
+    end
+
+    it 'should add new collections' do |ex|
+      cequel.schema.sync_table table_name do
+        key :blog_subdomain, :text
+        key :permalink, :text
+        column :title, :ascii, :index => true
+        column :body, :ascii
+        column :created_at, :timestamp
+        set :author_names, :text
+        with :comment, 'Test Table - #{ex.description}'
+
+        list :categories, :text # new
       end
 
-      it 'should add new collections' do
-        expect(table.column(:categories)).to be_a(Cequel::Schema::List)
+      expect(table.column(:categories)).to be_a(Cequel::Schema::List)
+    end
+
+    it 'should add new column with index' do  |ex|
+      cequel.schema.sync_table table_name do
+        key :blog_subdomain, :text
+        key :permalink, :text
+        column :title, :ascii, :index => true
+        column :body, :ascii
+        column :created_at, :timestamp
+        set :author_names, :text
+        with :comment, 'Test Table - #{ex.description}'
+
+        column :primary_author_id, :uuid, :index => true # new
       end
 
-      it 'should add new column with index' do
-        expect(table.column(:primary_author_id)).to be_indexed
+      expect(table.column(:primary_author_id)).to be_indexed
+    end
+
+    it 'should add index to existing columns' do |ex|
+      cequel.schema.sync_table table_name do
+        key :blog_subdomain, :text
+        key :permalink, :text
+        column :title, :ascii, :index => true
+        column :body, :ascii
+        set :author_names, :text
+        with :comment, 'Test Table - #{ex.description}'
+
+        column :created_at, :timestamp, index: true # altered
       end
 
-      it 'should add index to existing columns' do
-        expect(table.column(:created_at)).to be_indexed
+      expect(table.column(:created_at)).to be_indexed
+    end
+
+    it 'should drop index from existing columns' do |ex|
+      cequel.schema.sync_table table_name do
+        key :blog_subdomain, :text
+        key :permalink, :text
+        column :title, :ascii  # omitting index: true
+        column :body, :ascii
+        set :author_names, :text
+        with :comment, 'Test Table - #{ex.description}'
       end
 
-      it 'should drop index from existing columns' do
-        expect(table.column(:title)).not_to be_indexed
+      expect(table.column(:title)).not_to be_indexed
+    end
+
+    it 'should change properties' do |ex|
+      cequel.schema.sync_table table_name do
+        key :blog_subdomain, :text
+        key :permalink, :text
+        column :title, :ascii, :index => true
+        column :body, :ascii
+        column :created_at, :timestamp
+        set :author_names, :text
+        with :comment, 'Test Table 2.0' # altered
       end
 
-      it 'should change column type' do
-        expect(table.column(:body).type).to eq(Cequel::Type[:text])
-      end
-
-      it 'should change properties' do
-        expect(table.property(:comment)).to eq('Test Table 2.0')
-      end
-
+      expect(table.property(:comment)).to eq('Test Table 2.0')
     end
 
     context 'invalid migrations' do
@@ -107,6 +163,21 @@ describe Cequel::Schema::TableSynchronizer do
           end
         }.to raise_error(Cequel::InvalidSchemaMigration)
       end
+
+      it 'should change column type' do |ex|
+        expect{
+          cequel.schema.sync_table table_name do
+            key :blog_subdomain, :text
+            key :permalink, :text
+            column :title, :ascii, :index => true
+            column :created_at, :timestamp
+            set :author_names, :text
+            with :comment, 'Test Table - #{ex.description}'
+
+            column :body, :text # altered
+          end
+        }.to raise_error(Cequel::InvalidSchemaMigration)
+    end
 
       it 'should not allow adding a key' do
         expect {